requirements.txt
# Avatar Tank - Python Dependencies
# Core web framework
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.9.0
eventlet==0.33.3

# Computer vision and image processing
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# Network and HTTP requests
requests==2.31.0
urllib3==2.0.7

# Serial communication
pyserial==3.5

# Audio processing
pygame==2.5.2
pyaudio==0.2.11

# System utilities
psutil==5.9.6
python-dotenv==1.0.0

# JSON and data handling
jsonschema==4.19.2

# Date and time utilities
python-dateutil==2.8.2

# Logging and debugging
colorlog==6.8.0

# Optional: Development and testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

